+++ products/product.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
 class Product(ABC):
-    def __init__(self, name: str, price: float):
+    def __init__(self, name: str, price: int):
         self.id = None
         self.created_at = None
         self.name = name
